Add methods for setting templates via button lists
For relevant template types, convenience methods were added to template
classes to make setting layouts from lists of buttons easier. An
example use case for this would be having a list of buttons you want
to put into a small area on a ChestTemplate, so you would
use the squareFromList method to place them on the page.